requirements.txt
appnope==0.1.0
backcall==0.2.0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
decorator==4.4.2
Flask==1.1.2
flask-redis==0.4.0
idna==2.10
ipython==7.18.1
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.17.2
Jinja2==2.11.2
MarkupSafe==1.1.1
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.7
ptyprocess==0.6.0
Pygments==2.7.0
PyJWT==1.7.1
python-dotenv==0.14.0
pytz==2020.1
redis==3.5.3
requests==2.24.0
rq==1.5.1
six==1.15.0
traitlets==5.0.4
twilio==6.45.1
urllib3==1.25.10
wcwidth==0.2.5
Werkzeug==1.0.1
